+++ src/Enterspeed.Source.SitecoreCms.V8/Extensions/DateTimeExtensions.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Globalization;
+
+namespace Enterspeed.Source.SitecoreCms.V8.Extensions
+{
+    public static class DateTimeExtensions
+    {
+        public static string ToSqlDateTime(this DateTime dateTime)
+        {
+            return dateTime.ToString("yyyy-MM-dd HH:mm:ss", CultureInfo.InvariantCulture);
+        }
+    }
+}