+++ src/Enterspeed.Source.SitecoreCms.V8/Pipelines/Initialize/InitializeEnterspeed.cs
@@ -5,6 +5,7 @@ using Sitecore;
 using Sitecore.Abstractions;
 using Sitecore.Data;
 using Sitecore.Data.Items;
+using Sitecore.Data.Masters;
 using Sitecore.Globalization;
 using Sitecore.Pipelines;
 using Sitecore.SecurityModel;
@@ -18,8 +19,12 @@ namespace Enterspeed.Source.SitecoreCms.V8.Pipelines.Initialize
         private const string SiteIcon = "Applications/32x32/window_gear.png";
         private const string EnabledSitesHelpText = "Select the site items here with the same fullPath as the rootPath configured for the respective site(s).";
         private const string EnabledDictionariesHelpText = "Select the dictionary parent item, to push the item and all descendant dictionaries to Enterspeed.";
-
         private const string ApiKeyHelpText = "For example \"source-xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx\".";
+
+        private const string ApiBaseUrlTitle = "API Base Url *";
+        private const string ApiKeyTitle = "Api Key *";
+        private const string SiteBaseUrlTitle = "Site Base Url *";
+
 
         private readonly BaseItemManager _itemManager;
         private readonly BaseFactory _factory;
@@ -156,7 +161,7 @@ namespace Enterspeed.Source.SitecoreCms.V8.Pipelines.Initialize
                 ?? enterspeedSiteConfigTemplateItem.Add("Data", new TemplateID(TemplateIDs.TemplateSection), EnterspeedIDs.Templates.EnterspeedSiteConfigurationDataSectionID);
         }
 
-        private static void EnsureEnterspeedSiteConfigurationDataSectionFields(Item enterspeedSiteConfigSection)
+        private static void EnsureEnterspeedSiteConfigurationDataSectionFields(Item enterspeedSiteConfigSection, Database masterDb)
         {
             Item apiBaseUrlField = enterspeedSiteConfigSection.Children["API Base Url"] ?? enterspeedSiteConfigSection.Add("API Base Url", new TemplateID(TemplateIDs.TemplateField), EnterspeedIDs.Fields.EnterspeedApiBaseUrlFieldID);
 
@@ -180,6 +185,12 @@ namespace Enterspeed.Source.SitecoreCms.V8.Pipelines.Initialize
                 if (currentUnversionedValue != "1")
                 {
                     apiBaseUrlField[TemplateFieldIDs.Unversioned] = "1";
+                }
+
+                string currentTitle = apiBaseUrlField[TemplateFieldIDs.Title];
+                if (currentTitle != ApiBaseUrlTitle)
+                {
+                    apiBaseUrlField[TemplateFieldIDs.Title] = ApiBaseUrlTitle;
                 }
             }
 
@@ -212,6 +223,12 @@ namespace Enterspeed.Source.SitecoreCms.V8.Pipelines.Initialize
                 if (!currentHelpValue.Equals(ApiKeyHelpText, StringComparison.OrdinalIgnoreCase))
                 {
                     apiKeyField.Help.ToolTip = ApiKeyHelpText;
+                }
+
+                string currentTitle = apiKeyField[TemplateFieldIDs.Title];
+                if (currentTitle != ApiKeyTitle)
+                {
+                    apiKeyField[TemplateFieldIDs.Title] = ApiKeyTitle;
                 }
             }
 
@@ -307,6 +324,12 @@ namespace Enterspeed.Source.SitecoreCms.V8.Pipelines.Initialize
                 {
                     siteBaseUrlField[TemplateFieldIDs.Unversioned] = "1";
                 }
+
+                string currentTitle = siteBaseUrlField[TemplateFieldIDs.Title];
+                if (currentTitle != SiteBaseUrlTitle)
+                {
+                    siteBaseUrlField[TemplateFieldIDs.Title] = SiteBaseUrlTitle;
+                }
             }
 
             Item mediaBaseUrlField = enterspeedSiteConfigSection.Children["Media Base Url"]
@@ -387,6 +410,23 @@ namespace Enterspeed.Source.SitecoreCms.V8.Pipelines.Initialize
                 }
             }
 
+            var enterspeedSiteConfigTemplate = masterDb.GetTemplate(enterspeedSiteConfigSection.Parent.ID);
+            if (enterspeedSiteConfigTemplate.StandardValues == null)
+            {
+                using (new EditContext(enterspeedSiteConfigTemplate))
+                {
+                    enterspeedSiteConfigTemplate.CreateStandardValues();
+                }
+
+                using (new EditContext(enterspeedSiteConfigTemplate.StandardValues))
+                {
+                    var apiBaseUrlStandardValues = enterspeedSiteConfigTemplate.StandardValues["API Base Url"];
+                    if (apiBaseUrlStandardValues.ToString() == string.Empty)
+                    {
+                        enterspeedSiteConfigTemplate.StandardValues["API Base Url"] = "https://api.enterspeed.com";
+                    }
+                }
+            }
         }
 
         private void Init()
@@ -416,7 +456,7 @@ namespace Enterspeed.Source.SitecoreCms.V8.Pipelines.Initialize
             EnsureEnterspeedSiteConfigurationTemplateFields(enterspeedSiteConfigTemplateItem);
             Item enterspeedSiteConfigSection = EnsureEnterspeedSiteConfigurationTemplateDataSection(enterspeedSiteConfigTemplateItem);
 
-            EnsureEnterspeedSiteConfigurationDataSectionFields(enterspeedSiteConfigSection);
+            EnsureEnterspeedSiteConfigurationDataSectionFields(enterspeedSiteConfigSection, masterDb);
 
             Item enterspeedConfigTemplateItem = EnsureEnterspeedConfigurationTemplate(EnsureEnterspeedTemplatesFolder(templatesSystemRoot));
 
@@ -424,8 +464,6 @@ namespace Enterspeed.Source.SitecoreCms.V8.Pipelines.Initialize
             Item enterspeedConfigSection = EnsureEnterspeedConfigurationTemplateDataSection(enterspeedConfigTemplateItem);
 
             EnsureEnterspeedConfigurationDataSectionFields(enterspeedConfigSection);
-
-            _publishManager.PublishItem(templatesSystemRoot, new[] { webDb }, new[] { language }, true, false, true);
         }
 
         private void EnsureItems(Database masterDb, Database webDb, Language language)