+++ src/Enterspeed.Source.SitecoreCms.V9/Data/Models/EnterspeedContentState.cs
@@ -1,0 +1,8 @@
+﻿namespace Enterspeed.Source.SitecoreCms.V9.Data.Models
+{
+    public enum EnterspeedContentState
+    {
+        Publish,
+        Preview
+    }
+}