+++ src/Enterspeed.Source.SitecoreCms.V9/Services/Contracts/IEnterspeedConfigurationService.cs
@@ -1,10 +1,10 @@
 ﻿using System.Collections.Generic;
 using Enterspeed.Source.SitecoreCms.V9.Models.Configuration;
 
-namespace Enterspeed.Source.SitecoreCms.V9.Services
+namespace Enterspeed.Source.SitecoreCms.V9.Services.Contracts
 {
     public interface IEnterspeedConfigurationService
     {
-        List<EnterspeedSitecoreConfiguration> GetConfiguration();
+        List<EnterspeedSitecoreConfiguration> GetConfigurations();
     }
 }