+++ src/Enterspeed.Source.SitecoreCms.V9/Services/Contracts/IEnterspeedPropertyService.cs
@@ -3,7 +3,7 @@ using Enterspeed.Source.Sdk.Api.Models.Properties;
 using Enterspeed.Source.SitecoreCms.V9.Models.Configuration;
 using Sitecore.Data.Items;
 
-namespace Enterspeed.Source.SitecoreCms.V9.Services
+namespace Enterspeed.Source.SitecoreCms.V9.Services.Contracts
 {
     public interface IEnterspeedPropertyService
     {