+++ src/Enterspeed.Source.SitecoreCms.V9/Services/EnterspeedConfigurationService.cs
@@ -3,9 +3,11 @@ using System.Collections.Generic;
 using System.Linq;
 using Enterspeed.Source.SitecoreCms.V9.Exceptions;
 using Enterspeed.Source.SitecoreCms.V9.Models.Configuration;
+using Enterspeed.Source.SitecoreCms.V9.Services.Contracts;
 using Sitecore.Abstractions;
 using Sitecore.Data.Fields;
 using Sitecore.Data.Items;
+using Sitecore.Data.Managers;
 using Sitecore.Globalization;
 using Sitecore.Links;
 using Sitecore.Links.UrlBuilders;
@@ -47,7 +49,7 @@ namespace Enterspeed.Source.SitecoreCms.V9.Services
             _loggingService = loggingService;
         }
 
-        public List<EnterspeedSitecoreConfiguration> GetConfiguration()
+        public List<EnterspeedSitecoreConfiguration> GetConfigurations()
         {
             var enterspeedConfigurationItem = _itemManager.GetItem(EnterspeedIDs.Items.EnterspeedConfigurationID, Language.Parse("en"), Version.Latest, _factory.GetDatabase("web"));
             if (HasNoConfigurationSetUp(enterspeedConfigurationItem))
@@ -203,22 +205,28 @@ namespace Enterspeed.Source.SitecoreCms.V9.Services
 
         private bool IsConfigurationUpdated(Item item, out DateTime currentUpdatedDate)
         {
-            if (item.Children != null && item.Children.Any())
+            var database = _factory.GetDatabase("web");
+            var languages = _languageManager.GetLanguages(database);
+            currentUpdatedDate = DateTime.MinValue;
+
+            foreach (var language in languages)
             {
-                currentUpdatedDate = item.Children.Max(i => i.Statistics.Updated);
-            }
-            else
-            {
-                currentUpdatedDate = DateTime.UtcNow;
+                var languageItem = _itemManager.GetItem(item.ID, language, Version.Latest, database);
+                if (languageItem != null)
+                {
+                    if (languageItem.Children != null && languageItem.Children.Any())
+                    {
+                        currentUpdatedDate = languageItem.Children.Max(i => i.Statistics.Updated);
+                    }
+
+                    if (_lastUpdatedDate < currentUpdatedDate)
+                    {
+                        return true;
+                    }
+                }
             }
 
-            if (_lastUpdatedDate >= currentUpdatedDate &&
-                _configuration != null)
-            {
-                return false;
-            }
-
-            return true;
+            return false;
         }
     }
 }