+++ src/Enterspeed.Source.SitecoreCms.V9/Services/EnterspeedSitecoreFieldService.cs
@@ -1,4 +1,5 @@
-﻿using Sitecore.Data.Fields;
+﻿using Enterspeed.Source.SitecoreCms.V9.Services.Contracts;
+using Sitecore.Data.Fields;
 namespace Enterspeed.Source.SitecoreCms.V9.Services
 {
     public class EnterspeedSitecoreFieldService : IEnterspeedSitecoreFieldService